requirements.txt
mysql-connector-python>=8.0.32
pandas>=1.3.5
python-telegram-bot>=20.1
python-dotenv>=0.21.1
boto3>=1.28.50
